Provide better errors when deserializing `build.toml`
Before this change when there was an error in `build.toml`, one
would get a generic error:

```
Error: Cannot parse TOML in /Users/daniel/git/relu-metal/build.toml

Caused by:
    data did not match any variant of untagged enum BuildCompat
```

However, what we want is to give the error when trying to parse as the
V2 format, since the V1 format is only there for backwards
compat/updates. With this change the error of parsing as V2 is bubbled
up. For instance:

```
Error: Cannot parse TOML in /Users/daniel/git/relu-metal/build.toml

Caused by:
    Unknown variant metal. Expected one of cuda, rocm
```