Add a full ReLU example with backprop and `torch.compile` support
The existing `examples/relu` is made to be as simple as possible,
using the absolute minimum to get a CUDA/ROCm/XPU kernel working.
However, real-world kernels often also have to provide backprop and
`torch.compile` support.

This change adds an example of an extended ReLU kernel that can
backpropagate and can be compiled with `torch.compile`.