Fix a regression in test shells
The XPU support PR changed stdenv selection to choose the default stdenv
when one of the kernels is an XPU kernel. However, this does not work
for mixed XPU/CUDA projects, since we need to select the CUDA backend
stdenv when building for CUDA.

Change the logic to always use the stdenv for the backend that we are
building for.