Overlay Torch in the package set
Before this change, we wouldn't overlay Torch in the package set. The
main reason was that evaluation was too expensive -- even more so when
we also had C++98 and C++11 variants (which doubled the number of Torch
versions).

However, this does not really work well with dependencies (some
dependencies like `einops` will rely on Torch) or extra test
dependencies that a user adds to a flake. So overlay Torch so that all
of nixpkgs sees the version that is defined in the build set.

In this change, the `cxx11Abi` is also removed and the C++11 ABI is set
as the default. Upstream does not build for the C++98 ABI anymore, so it
is time to remove the support before it bitrots.