Separate build data structs from serializations
We have three versions of the build format (v1-3), where we converted
prior versions to the latest and used the latest as the actual
configuration. The downside of this approach is that every time we
introduce a new version, we have to move all utility functions, etc.
to the latest.

This change decouples the configuration that is used to generate the
build files from the concrete serializations. The core data structures
move to the top-level module and v1-v3.rs now only contain plain data
structures to (de)serialize with serde.